requirements.txt
anyio==3.5.0
asgiref==3.5.0
backtrader==1.9.76.123
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.0.3
colorama==0.4.4
cryptography==37.0.4
et-xmlfile==1.1.0
fastapi==0.73.0
Flask==2.2.2
h11==0.13.0
idna==3.3
investpy==1.0.8
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
numpy==1.22.1
oauthlib==3.2.0
openpyxl==3.0.10
pandas==1.4.0
pycparser==2.21
pydantic==1.9.0
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.29
sseclient
starlette==0.17.1
td-ameritrade-python-api==0.3.4
typing_extensions==4.0.1
Unidecode==1.3.4
urllib3==1.26.11
uvicorn==0.17.0.post1
websocket-client==1.3.2
websockets==10.3
Werkzeug==2.2.2
